src/app/components/document/document.service.ts
import { Injectable } from '@angular/core';
import { Http, Response } from '@angular/http';
import { Observable } from 'rxjs/Observable';

@Injectable()
export class DocumentService {

  private api = 'http://localhost:8080/api/';

  constructor(private http: Http) { }

  getDocument(name: string) {
    return this.http.get(this.api + 'parse/' + name)
      .map(this.extractData)
      .catch(this.handleError);
  }

  getDocuments() {
    return this.http.get(this.api + 'documents')
      .map(this.extractData)
      .catch(this.handleError);
  }

  private extractData(res: Response) {
    let body = res.json();
    return body.data || {};
  }
  private handleError(error: Response | any) {
    // In a real world app, you might use a remote logging infrastructure
    let errMsg: string;
    if (error instanceof Response) {
      const body = error.json() || '';
      const err = body.error || JSON.stringify(body);
      errMsg = `${error.status} - ${error.statusText || ''} ${err}`;
    } else {
      errMsg = error.message ? error.message : error.toString();
    }
    console.error(errMsg);
    return Observable.throw(errMsg);
  }

}
